fix: Phase 10 - Fix optimization module mypy type annotations
- backtest_presenter.py: Fix import order (future annotations first)
- base_optimizer.py: Add float() cast for np.mean, add additional_info field
- bayesian_optimizer.py: Add List[TrialResult] type annotation, rename min_val/max_val
- grid_search_optimizer.py: Fix type annotations for float variables
- momentum_auto_optimizer.py: Minor type fix
- parameter/base_optimizer.py: Fix type annotations

All files now pass 11/11 validation checks.

+++ app/domain/optimization/momentum_auto_optimizer.py
@@ -227,7 +227,7 @@ class MomentumAutoOptimizer:
             if not scores:
                 return -1000.0
 
-            return np.mean(scores)
+            return float(np.mean(scores))
 
         # Run optimization
         study.optimize(objective, n_trials=20)  # Limited trials for monthly updates
